integration-tests/requirements.txt
behave
black
environs
isort
mypy
python-dateutil
python-jose==3.*
reportportal-behave-client==1.*
requests
she_logging
types-python-dateutil
types-requests
